requirements.txt
streamlit==1.29.0
pandas==2.1.3
numpy==1.26.1
langchain==0.0.348
langchain-community==0.0.16
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-docx==1.0.1
pdfplumber==0.10.3
bcrypt==4.0.1
sqlalchemy==2.0.23
supabase==2.0.3
chromadb==0.4.22
pymupdf==1.23.7